+++ jackson-datatype-nostr/src/main/java/org/tbk/jackson/datatype/nostr/NostrModule.java
@@ -7,6 +7,7 @@ import com.fasterxml.jackson.databind.module.SimpleSerializers;
 import org.tbk.nostr.proto.Event;
 import org.tbk.nostr.proto.ProfileMetadata;
 import org.tbk.nostr.proto.Request;
+import org.tbk.nostr.proto.Response;
 
 public class NostrModule extends Module {
     @Override
@@ -24,12 +25,14 @@ public class NostrModule extends Module {
         SimpleSerializers serializers = new SimpleSerializers();
         serializers.addSerializer(new EventSerializer());
         serializers.addSerializer(new RequestSerializer());
+        serializers.addSerializer(new ResponseSerializer());
         serializers.addSerializer(new ProfileMetadataSerializer());
         context.addSerializers(serializers);
 
         SimpleDeserializers deserializers = new SimpleDeserializers();
         deserializers.addDeserializer(Event.class, new EventDeserializer());
         deserializers.addDeserializer(Request.class, new RequestDeserializer());
+        deserializers.addDeserializer(Response.class, new ResponseDeserializer());
         deserializers.addDeserializer(ProfileMetadata.class, new ProfileMetadataDeserializer());
         context.addDeserializers(deserializers);
     }